Refactor java+js to use eager marker custom elem
"wc-ajax-eager" elements are now used to mark eager containers.
Currently they are simple divs that sit as a child of the eager
containers, with that container's id as its inner html. This is how ids
are fed into the ajax register.

Currently, since the ids are given as text inside the div, users can
actually see that container id for the few seconds that the ajax load
is happening. This should be fixed in a future commit.