Comment out eager-spotting xslt renderer for panel
Previous xslt did not render the contents of eager panels. This is a
problem for the current implementation since the html custom element
that acts as a marker for eager panels is a child element of that panel.
That code is commented out for now; will have to ask best way to deal
with that/whether that check is necessary.

Other eager containers seem to be working fine with new implementation
EXCEPT for figures and sections. Will have to investigate.