Change custom rtf element to extend textarea
Instead of using a blank custom element to mark rtf fields, the actual
input textarea element is extended as a custom element. This is to
handle the rtf import problem, since now it is easy to detect when the
rtf module should be imported through the existance of the textarea in
textarea.mjs.

Potential problem: this functionality only plays nice with rtf
components that are not read-only. This is not a problem if components
are unable to switch between being read-only and not, since read-only
rtf fields don't seem to have a need to be registered through TinyMCE.
If that is not the case, however, it will have to be accounted for in
another way.