Switch back to using non-extending custom element
Due to webkit not supporting the extension of existing html elements
custom ones, regular custom elements have to be used. It is back to
using the empty wc-rtf marker element.

Note that the definition of this element is still done in textArea.mjs,
so rtf.mjs can continue being imported on-demnad.