Add better unit tests for rtf
Using a mock of TinyMCE, rtf is now tested for default configuration.

Slight remodel of the rtf code to use resourceLoader to import TinyMCE
was also done. This was to make it possible to mock TinyMCE by pointing
the testing suite towards the mock with an import map in the DOM.